requirements.txt
# torch==2.2.2
# torchvision==0.17.2  скачиваются в dockerfile, без GPU
pandas==2.1.4
scikit-learn==1.3.2
pillow==10.0.1
matplotlib==3.8.2
tqdm==4.66.1
streamlit==1.50.0
plotly==5.17.0